added property comparison triggers and updated property files

+++ src/authoring/model/triggers/selfconditions/PropertyLessThan.java
@@ -8,7 +8,7 @@ public abstract class PropertyLessThan extends APropertyTest {
 	 */
 	private static final long serialVersionUID = -7330673705733755062L;
 
-	protected boolean checkCondition(int property, Double value) {
+	protected boolean checkCondition(Double property, Double value) {
 		return Double.compare(property, value) < 0;
 	}
 }
